Fix API Overload Issue
There was an issue with the friends list that caused the Steam API to timeout. I worked around this by hiding the friends list rather than unmounting it when a user selected a friend, I also added a cooldown to the API requests in the server when fetching friends to further prevent overload.